css/selectors/has-relative-argument.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>:has pseudo class behavior with various relative arguments</title>
<link rel="author" title="Byungwoo Lee" href="mailto:blee@igalia.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/selectors/#relational">

<main id=main>
 <div id=d01>
  <div id=d02 class="scope">
    <div id=d03 class="a"></div>
    <div id=d04></div>
    <div id=d05 class="b"></div>
  </div>
  <div id=d06 class="scope">
    <div id=d07 class="scope">
      <div id=d08 class="a"></div>
    </div>
  </div>
  <div id=d09 class="scope">
    <div id=d10 class="a">
      <div id=d11 class="b"></div>
    </div>
  </div>
  <div id=d12 class="scope">
    <div id=d13 class="a">
      <div id=d14>
        <div id=d15 class="b"></div>
      </div>
    </div>
    <div id=d16 class="b"></div>
  </div>
 </div>
 <div id=d17>
  <div id=d18 class="scope"></div>
  <div id=d19 class="scope"></div>
  <div id=d20 class="a"></div>
  <div id=d21 class="scope"></div>
  <div id=d22 class="a">
   <div id=d23 class="b"></div>
  </div>
  <div id=d24 class="scope"></div>
  <div id=d25 class="a">
   <div id=d26>
    <div id=d27 class="b"></div>
   </div>
  </div>
  <div id=d28 class="scope"></div>
  <div id=d29 class="a"></div>
  <div id=d30 class="b">
   <div id=d31 class="c"></div>
  </div>
  <div id=d32 class="scope"></div>
  <div id=d33 class="a"></div>
  <div id=d34 class="b">
   <div id=d35>
    <div id=d36 class="c"></div>
   </div>
  </div>
  <div id=d37 class="scope"></div>
  <div id=d38 class="a"></div>
  <div id=d39 class="b"></div>
  <div id=d40 class="scope"></div>
  <div id=d41 class="a"></div>
  <div id=d42></div>
  <div id=d43 class="b">
   <div id=d44 class="scope">
    <div id=d45 class="c"></div>
   </div>
  </div>
  <div id=d46 class="scope"></div>
  <div id=d47 class="a">
  </div>
 </div>
 <div>
  <div id=d48 class="scope">
   <div id=d49 class="scope">
    <div id=d50 class="scope d">
     <div id=d51 class="scope d">
      <div id=d52 class="scope">
       <div id=d53 class="scope e">
        <div id=d54 class="f"></div>
       </div>
      </div>
     </div>
    </div>
   </div>
  </div>
  <div id=d55 class="scope"></div>
  <div id=d56 class="scope d"></div>
  <div id=d57 class="scope d"></div>
  <div id=d58 class="scope"></div>
  <div id=d59 class="scope e"></div>
  <div id=d60 class="f"></div>
 </div>
 <div>
  <div id=d61 class="scope1"></div>
  <div id=d62 class="scope1 scope2"></div>
  <div id=d63 class="scope1 scope2">
   <div id=d64 class="scope2 g">
    <div id=d65 class="scope2">
     <div id=d66 class="scope2 h">
      <div id=d67 class="i"></div>
     </div>
    </div>
   </div>
  </div>
  <div id=d68 class="scope1 scope2">
   <div id=d69 class="scope1"></div>
   <div id=d70 class="scope1"></div>
   <div id=d71 class="scope1 scope2">
    <div id=d72 class="scope2 g">
     <div id=d73 class="scope2">
      <div id=d74 class="scope2 h">
       <div id=d75 class="i"></div>
      </div>
     </div>
    </div>
   </div>
   <div id=d76 class="scope1"></div>
   <div id=d77 class="j"><div id=d78><div id=d79></div></div></div>
  </div>
  <div id=d80 class="j"></div>
 </div>
</main>

<script>
  function formatElements(elements) {
    return elements.map(e => e.id).sort().join();
  }

  // Test that |selector| returns the given elements in #main.
  function test_selector_all(selector, expected) {
    test(function() {
      let actual = Array.from(main.querySelectorAll(selector));
      assert_equals(formatElements(actual), formatElements(expected));
    }, `${selector} matches expected elements`);
  }

  test_selector_all('.scope:has(:scope > .a)', [d02, d07, d09, d12]);
  test_selector_all('.scope:has(:scope > .a > .b)', [d09]);
  test_selector_all('.scope:has(:scope > .a .b)', [d09, d12]);
  test_selector_all('.scope:has(:scope > .a + .b)', [d12]);
  test_selector_all('.scope:has(:scope > .a ~ .b)', [d02, d12]);

  test_selector_all('.scope:has(:scope + .a)',
                    [d19, d21, d24, d28, d32, d37, d40, d46]);
  test_selector_all('.scope:has(:scope + .a > .b)', [d21]);
  test_selector_all('.scope:has(:scope + .a .b)', [d21, d24]);
  test_selector_all('.scope:has(:scope + .a + .b)', [d28, d32, d37]);
  test_selector_all('.scope:has(:scope + .a ~ .b)',
                    [d19, d21, d24, d28, d32, d37, d40]);

  test_selector_all('.scope:has(:scope ~ .a)',
                    [d18, d19, d21, d24, d28, d32, d37, d40, d46]);
  test_selector_all('.scope:has(:scope ~ .a > .b)',
                    [d18, d19, d21]);
  test_selector_all('.scope:has(:scope ~ .a .b)',
                    [d18, d19, d21, d24]);
  test_selector_all('.scope:has(:scope ~ .a + .b)',
                    [d18, d19, d21, d24, d28, d32, d37]);
  test_selector_all('.scope:has(:scope ~ .a + .b > .c)',
                    [d18, d19, d21, d24, d28]);
  test_selector_all('.scope:has(:scope ~ .a + .b .c)',
                    [d18, d19, d21, d24, d28, d32]);

  test_selector_all('.scope:has(.d .e)', [d48, d49, d50]);
  test_selector_all('.scope:has(.d .e) .f', [d54]);
  test_selector_all('.scope:has(:scope .d .e)', [d48, d49, d50]);
  test_selector_all('.scope:has(:scope .d .e) .f', [d54]);
  test_selector_all('.scope:has(:scope > .d)', [d49, d50]);
  test_selector_all('.scope:has(:scope > .d) .f', [d54]);
  test_selector_all('.scope:has(:scope ~ .d ~ .e)', [d48, d55, d56]);
  test_selector_all('.scope:has(:scope ~ .d ~ .e) ~ .f', [d60]);
  test_selector_all('.scope:has(:scope + .d ~ .e)', [d55, d56]);
  test_selector_all('.scope:has(:scope + .d ~ .e) ~ .f', [d60]);

  test_selector_all('.scope2:has(:scope > .g .h)', [d63, d71])
  test_selector_all('.scope2:has(:scope .g .h)', [d63, d68, d71])
  test_selector_all('.scope2:has(:scope > .g .h) .i', [d67, d75])
  test_selector_all('.scope2:has(:scope .g .h) .i', [d67, d75])
  test_selector_all('.scope1:has(:scope + .scope2:has(:scope > .g .h) .i)',
      [d62, d70])
  test_selector_all('.scope1:has(:scope + .scope2:has(:scope .g .h) .i)',
      [d62, d63, d70])
  test_selector_all(
      '.scope1:has(:scope + .scope2:has(:scope > .g .h) .i) ~ .j',
      [d77, d80])
  test_selector_all(
      '.scope1:has(:scope + .scope2:has(:scope .g .h) .i) ~ .j',
      [d77, d80])
  test_selector_all('.scope1:has(:scope ~ .scope2:has(:scope > .g .h) .i)',
      [d61, d62, d69, d70])
  test_selector_all('.scope1:has(:scope ~ .scope2:has(:scope .g .h) .i)',
      [d61, d62, d63, d69, d70])
</script>
